style(package): remove underscore in package name
