+++ src/main/java/com/e2i/wemeet/domain/profileimage/ProfileImage.java
@@ -1,4 +1,4 @@
-package com.e2i.wemeet.domain.profile_image;
+package com.e2i.wemeet.domain.profileimage;
 
 import com.e2i.wemeet.domain.base.BaseTimeEntity;
 import com.e2i.wemeet.domain.member.Member;
